src/requirements.txt
geopandas==1.1.3
joblib==1.5.2
matplotlib==3.10.8
numpy==2.4.3
optuna==4.7.0
pandas==3.0.1
PyYAML==6.0.3
rasterio==1.4.4
scikit_learn==1.8.0
Shapely==2.1.2
